intent-trainer/requirements.txt
# Intent classifier training (fastText). Python 3.8+
# 预编译 wheel，避免官方 fasttext 的 C++ 编译。API 与 fasttext 一致。
fasttext-wheel>=0.9.2
